fix variable cast

+++ src/AuctionHouseBot.cpp
@@ -1891,7 +1891,7 @@ std::pair<uint64, uint64> AuctionHouseBot::CalculateMovingAveragePrices(uint32 i
         do
         {
             Field* fields = result->Fetch();
-            uint64 finalPrice = fields[0].GetUInt64();
+            uint64 finalPrice = static_cast<uint64>(fields[0].GetUInt32());
             std::string auctionType = fields[1].GetString();
 
             if (auctionType == "buyout")
@@ -1963,8 +1963,8 @@ std::pair<uint64, uint64> AuctionHouseBot::CalculateMovingAveragePrices(uint32 i
     if (priceOverrideResult)
     {
         Field* fields = priceOverrideResult->Fetch();
-        uint64 avgPrice = fields[0].GetUInt64();
-        uint64 minPrice = fields[1].GetUInt64();
+        uint64 avgPrice = static_cast<uint64>(fields[0].GetUInt32());
+        uint64 minPrice = static_cast<uint64>(fields[1].GetUInt32());
 
         // Enforce minimum and maximum price limits based on avgPrice and minPrice values
         uint64 maxPrice = avgPrice * 2; // Example: max price is twice the average price
@@ -1998,8 +1998,8 @@ void AuctionHouseBot::AdjustPrices(uint32 itemId, uint64& buyoutPrice, uint64& b
     if (priceOverrideResult)
     {
         Field* fields = priceOverrideResult->Fetch();
-        uint64 avgPrice = fields[0].GetUInt64();
-        uint64 minPrice = fields[1].GetUInt64();
+        uint64 avgPrice = static_cast<uint64>(fields[0].GetUInt32());
+        uint64 minPrice = static_cast<uint64>(fields[1].GetUInt32());
 
         uint64 maxPrice = avgPrice * 2; // Example: max price is twice the average price
         uint64 adjustedMinPrice = minPrice * config->MinPriceTolerance; // Allow prices to go slightly below minPrice
